[Chore] #44 - 프로젝트 디렉토리 구조화
- 프로젝트의 디렉토리를 변경하였습니다.

- 누락된 접근제어자를 추가하고, 미사용하는 코드를 삭제하였습니다.
