+++ WeatherWhat/WeatherWhat/MainView/MainViewModel.swift
@@ -8,16 +8,17 @@
 import Foundation
 import RxSwift
 import RxCocoa
+import RxMoya
 import Moya
 
 final class MainViewModel {
-    
+
     private var disposeBag = DisposeBag()
     private let weatehrProvider = MoyaProvider<WeatherAPI>()
     private let errorRelay = PublishRelay<Error>()
-    
+
     func transform(_ input: Input) -> Output {
-        
+
         let location = input.fetchInitialData
             .withUnretained(self)
             .flatMap { vm, _ -> Observable<LocationData> in
@@ -28,22 +29,21 @@ final class MainViewModel {
                         return .empty()
                     }
             }
-        
+
         let networkResponse = location
             .withUnretained(self)
             .flatMap { vm, location -> Observable<(CurrentWeather, ForecastWeather)> in
                 let currentWeather = vm.fetchCurrentWeather(lat: location.lat,
                                                             lon: location.lon)
                 let forecastWeather = vm.fetchForecastWeatehr(lat: location.lat,
-                                                             lon: location.lon)
-                
+                                                              lon: location.lon)
+
                 return Observable.zip(currentWeather, forecastWeather)
             }
             .share()
-        
+
         let collectionViewData = Observable.merge(input.fetchInitialData,
                                                   input.toggleButtonTapped.asObservable())
-            .debug()
             .flatMapLatest { isCelsius in
                 return networkResponse.map {
                     ($0.0, $0.1, isCelsius)
@@ -54,64 +54,64 @@ final class MainViewModel {
             }
             .withUnretained(self)
             .map { vm, response -> [SectionOfCellModel] in
-                
+
                 let currenWeatehrData = response.0
                 let forecastWeatherData = response.1
                 let isCelsius = response.2
                 let currentLocation = response.3
-                
+
                 let currentWeatherSectionItem = vm.createFirstSectionItem(
                     weather: currenWeatehrData,
                     location: currentLocation,
                     isCelsius: isCelsius
                 )
-                
+
                 let timeForecastSectionItem = vm.createSecondSectionItem(
                     weather: forecastWeatherData,
                     isCelsius: isCelsius
                 )
-                
+
                 let rainPercentSectionItem = vm.createThirdSectionItem(
                     weather: forecastWeatherData,
                     isCelsius: isCelsius
                 )
-                
+
                 let dayForecastSectionItem = vm.createForthSectionItem(
                     with: forecastWeatherData,
                     isCelsius: isCelsius
                 )
-                
+
                 let currentWeatherSection = SectionOfCellModel(
                     section: .currentWeather,
                     items: currentWeatherSectionItem
                 )
-                
+
                 let timeForecastSection = SectionOfCellModel(
                     section: .timeForecastCellModel,
                     items: timeForecastSectionItem
                 )
-                
+
                 let rainPercentSection = SectionOfCellModel(
                     section: .rainPercentResult,
                     items: rainPercentSectionItem
                 )
-                
+
                 let dayForecastSection = SectionOfCellModel(
                     section: .dayForecastResult,
                     items: dayForecastSectionItem
                 )
-                
+
                 let dataSource: [SectionOfCellModel] = [
                     currentWeatherSection,
                     timeForecastSection,
                     rainPercentSection,
                     dayForecastSection
                 ]
-                
+
                 return dataSource
             }
             .asDriver(onErrorDriveWith: .empty())
-        
+
         let background = input.fetchInitialData
             .flatMapFirst { _ in
                 return networkResponse.map {
@@ -120,14 +120,14 @@ final class MainViewModel {
             }
             .take(1)
             .asDriver(onErrorDriveWith: .empty())
-        
+
         return Output(
             collectionViewData: collectionViewData,
             backgorund: background,
             error: errorRelay.asDriver(onErrorDriveWith: .empty())
         )
     }
-    
+
     private func getCurrentLocation() ->  Observable<LocationData> {
         return Observable<LocationData>.create { observer in
             do {
@@ -139,19 +139,19 @@ final class MainViewModel {
             return Disposables.create()
         }
     }
-    
+
     private func fetchCurrentWeather(lat: String, lon: String) -> Observable<CurrentWeather> {
         return weatehrProvider.rx.request(.currentCelsius(lat: lat, lon: lon))
             .asObservable()
             .map(CurrentWeather.self)
     }
-    
+
     private func fetchForecastWeatehr(lat: String, lon: String) -> Observable<ForecastWeather> {
         return weatehrProvider.rx.request(.forecastCelsius(lat: lat, lon: lon))
             .asObservable()
             .map(ForecastWeather.self)
     }
-    
+
     private func setUnits(number: Double, isCelsius: Bool) -> String {
         let temperature = Measurement(value: number, unit: UnitTemperature.celsius)
         switch isCelsius {
@@ -161,7 +161,7 @@ final class MainViewModel {
             return "\(Int(temperature.converted(to: .fahrenheit).value))°"
         }
     }
-    
+
     private func createFirstSectionItem(weather data: CurrentWeather,
                                         location: LocationData,
                                         isCelsius: Bool) -> [SectionOfCellModel.Item] {
@@ -179,11 +179,11 @@ final class MainViewModel {
                 description: data.weather.first!.description.displayName, backgroundColor: sendWeatherColor(with: data)))
         ]
     }
-    
-    
+
+
     private func createSecondSectionItem(weather data: ForecastWeather,
                                          isCelsius: Bool) -> [SectionOfCellModel.Item] {
-        
+
         return data.list.map { forecastData in
             SectionOfCellModel.Item.timeForecastCellModel(
                 .init(
@@ -195,10 +195,10 @@ final class MainViewModel {
             )
         }
     }
-    
+
     private func createThirdSectionItem(weather data: ForecastWeather,
                                         isCelsius: Bool) -> [SectionOfCellModel.Item] {
-        
+
         return data.list.map { forecastData in
             SectionOfCellModel.Item.rainPercentResult(
                 .init(
@@ -208,22 +208,22 @@ final class MainViewModel {
             )
         }
     }
-    
+
     private func createForthSectionItem(with data: ForecastWeather,                     isCelsius: Bool) -> [SectionOfCellModel.Item] {
         let convertedForecasData = convetForecastData(with: data)
-        
+
         let dic = Dictionary(
             grouping: convertedForecasData,
             by: { $0.dtTxt }
         )
-        
+
         let convertedData: [SectionOfCellModel.CellModel] = dic
             .sorted(by: {
                 $0.key < $1.key
             })
             .compactMap { (date, data) in
                 let icons = data.flatMap { $0.weather.map { $0.icon} }
-                
+
                 let iconGroup = Dictionary(grouping: icons,
                                            by: {$0}
                 )
@@ -240,19 +240,19 @@ final class MainViewModel {
                                                )
                 )
             }
-        
+
         return convertedData
     }
-    
+
     private func sendWeatherColor(with data: CurrentWeather) -> String {
         let defaultColorName = "pureWhite"
         let weatherColor = WeatherIconMatchModel()
-        
+
         guard let list = data.weather.first else { return defaultColorName }
         guard let colorName = weatherColor.dictionary[list.icon] else { return defaultColorName }
         return colorName
     }
-    
+
     private func convetForecastData(with data: ForecastWeather) -> [ForecastWeather.List] {
         return data.list.map { forecast in
             let weather = [
@@ -261,7 +261,7 @@ final class MainViewModel {
                     icon: String(forecast.weather.first!.icon.prefix(2))
                 )
             ]
-            
+
             let day = convertDateTimeToDate(input: forecast.dtTxt)
             return ForecastWeather.List(
                 main: forecast.main,
@@ -271,50 +271,50 @@ final class MainViewModel {
             )
         }
     }
-    
+
     private func convertTimeToData(input: String) -> String {
         let dateFormatter = DateFormatter()
         let dateTimeFormatter = DateFormatter()
         dateFormatter.dateFormat = "yyyy-MM-dd HH:mm:ss"
         dateTimeFormatter.dateFormat = "HH시"
-        
+
         guard let dateData = dateFormatter.date(from: input) else {
             return ""
         }
-        
+
         return dateTimeFormatter.string(from: dateData)
     }
-    
+
     private func convertDateTimeToDate(input: String) -> String {
         let dateFormatter = DateFormatter()
         let dateFilterFormatter = DateFormatter()
-        
+
         dateFormatter.dateFormat = "yyyy-MM-dd HH:mm:ss"
         dateFilterFormatter.dateFormat = "yyyy-MM-dd"
-        
+
         guard let dateData = dateFormatter.date(from: input) else {
             return ""
         }
-        
+
         return dateFilterFormatter.string(from: dateData)
     }
-    
+
     private func convertDateToDay(input: String) -> String {
-        
+
         let dateDayFormatter = DateFormatter()
         let dateFilterFormatter = DateFormatter()
-        
+
         dateDayFormatter.dateFormat = "E"
         dateFilterFormatter.dateFormat = "yyyy-MM-dd"
         dateDayFormatter.locale = Locale(identifier: "ko_KR")
-        
+
         guard let dateData = dateFilterFormatter.date(from: input) else {
             return ""
         }
-        
+
         return dateDayFormatter.string(from: dateData)
     }
-    
+
 }
 
 extension MainViewModel {
@@ -322,7 +322,7 @@ extension MainViewModel {
         let fetchInitialData: Observable<Bool>
         let toggleButtonTapped: PublishRelay<Bool>
     }
-    
+
     struct Output {
         let collectionViewData: Driver<[SectionOfCellModel]>
         let backgorund: Driver<String>
