+++ WeatherWhat/WeatherWhat/MainView/Model/CurrentWeather.swift
