+++ WeatherWhat/WeatherWhat/MainView/Model/LocationData.swift
