+++ WeatherWhat/WeatherWhat/MainView/Model/SectionOfCellModel.swift
