+++ WeatherWhat/WeatherWhat/SearchView/Cell/LocationHistoryCell.swift
