--- WeatherWhat/WeatherWhat/View/CustomSearchBar.swift
@@ -1,8 +1,0 @@
-//
-//  CustomSearchBar.swift
-//  WeatherWhat
-//
-//  Created by Lee on 5/27/25.
-//
-
-import Foundation
