--- WeatherWhat/WeatherWhat/View/ViewController.swift
@@ -1,19 +1,0 @@
-//
-//  ViewController.swift
-//  WeatherWhat
-//
-//  Created by LCH on 5/21/25.
-//
-
-import UIKit
-import RxMoya
-import Moya
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-    }
-
-}
-
